:sparkles: New LowerCaseTypeDeclaration sniff
New sniff to verify that all non-classname based type declarations are lowercase, both for parameter type declarations as well as return type declarations.

Includes extensive unit tests.
Includes auto-fixer.

Added to the `Extra` ruleset.

This sniff is a duplicate of the same sniff as pulled upstream and can be removed once the minimum PHPCS requirement of WPCS has gone up beyond the PHPCS version in which that PR will be merged.

See: squizlabs/PHP_CodeSniffer 1685

This sniff has a slight overlap with PR 1084. I would suggest for PR 1084 to be adjusted to remove the overlap so the upstream version of this sniff can be used as a one-on-one replacement later on.

Loosely related to 764